feat: created reusable logo component
